Operation storage moved to core components
