+++ src/neptune/internal/backends/neptune_backend.py
@@ -26,6 +26,7 @@ from typing import (
 
 from neptune.api.dtos import FileEntry
 from neptune.common.exceptions import NeptuneException
+from neptune.core.components.operation_storage import OperationStorage
 from neptune.internal.artifacts.types import ArtifactFileData
 from neptune.internal.backends.api_model import (
     ApiExperiment,
@@ -55,7 +56,6 @@ from neptune.internal.id_formats import (
     UniqueId,
 )
 from neptune.internal.operation import Operation
-from neptune.internal.operation_processors.operation_storage import OperationStorage
 from neptune.internal.utils.git import GitInfo
 from neptune.internal.websockets.websockets_factory import WebsocketsFactory
