+++ src/neptune/internal/operation_processors/async_operation_processor.py
@@ -40,13 +40,13 @@ from neptune.common.warnings import (
 )
 from neptune.constants import ASYNC_DIRECTORY
 from neptune.core.components.metadata_file import MetadataFile
+from neptune.core.components.operation_storage import OperationStorage
 from neptune.envs import NEPTUNE_SYNC_AFTER_STOP_TIMEOUT
 from neptune.exceptions import NeptuneSynchronizationAlreadyStoppedException
 from neptune.internal.disk_queue import DiskQueue
 from neptune.internal.init.parameters import DEFAULT_STOP_TIMEOUT
 from neptune.internal.operation import Operation
 from neptune.internal.operation_processors.operation_processor import OperationProcessor
-from neptune.internal.operation_processors.operation_storage import OperationStorage
 from neptune.internal.operation_processors.utils import (
     common_metadata,
     get_container_dir,
