+++ src/neptune/internal/operation_processors/offline_operation_processor.py
@@ -25,10 +25,10 @@ from typing import (
 
 from neptune.constants import OFFLINE_DIRECTORY
 from neptune.core.components.metadata_file import MetadataFile
+from neptune.core.components.operation_storage import OperationStorage
 from neptune.internal.disk_queue import DiskQueue
 from neptune.internal.operation import Operation
 from neptune.internal.operation_processors.operation_processor import OperationProcessor
-from neptune.internal.operation_processors.operation_storage import OperationStorage
 from neptune.internal.operation_processors.utils import (
     common_metadata,
     get_container_dir,
