+++ src/neptune/internal/operation_processors/sync_operation_processor.py
@@ -24,8 +24,8 @@ from typing import (
 
 from neptune.constants import SYNC_DIRECTORY
 from neptune.core.components.metadata_file import MetadataFile
+from neptune.core.components.operation_storage import OperationStorage
 from neptune.internal.operation_processors.operation_processor import OperationProcessor
-from neptune.internal.operation_processors.operation_storage import OperationStorage
 from neptune.internal.operation_processors.utils import (
     common_metadata,
     get_container_dir,
