+++ tests/unit/neptune/new/client/test_run.py
@@ -110,7 +110,7 @@ class TestClientRun(AbstractExperimentTestMixin, unittest.TestCase):
         new=lambda path: os.path.normpath(os.path.join("/home/user/main_dir", path)),
     )
     @unittest.skipIf(IS_WINDOWS, "Linux/Mac test")
-    @patch("neptune.internal.operation_processors.operation_storage.os.listdir", new=lambda path: [])
+    @patch("neptune.core.components.operation_storage.os.listdir", new=lambda path: [])
     @patch("neptune.core.components.metadata_file.open", mock_open())
     def test_entrypoint(self):
         with init_run(mode="debug") as exp:
@@ -200,7 +200,7 @@ class TestClientRun(AbstractExperimentTestMixin, unittest.TestCase):
         "neptune.internal.utils.os.path.abspath",
         new=lambda path: os.path.normpath(os.path.join("/home/user/main_dir", path)),
     )
-    @patch("neptune.internal.operation_processors.operation_storage.os.listdir", new=lambda path: [])
+    @patch("neptune.core.components.operation_storage.os.listdir", new=lambda path: [])
     @patch("neptune.core.components.metadata_file.open", mock_open())
     def test_entrypoint_without_common_root(self):
         with init_run(mode="debug", source_files=["../*"]) as exp:
